% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi-impute.R
\name{epi_impute}
\alias{epi_impute}
\title{Epi-Impute: single-cell RNA-seq imputation via integration with single-cell ATAC-seq}
\usage{
epi_impute(
  sc_exp_data,
  sc_atac_data,
  sc_atac_cell_names,
  sc_atac_peaks_ann,
  cell_types,
  atac_bin_thrld = 100,
  ...
)
}
\arguments{
\item{sc_exp_data}{a single-cell RNA-seq count matrix, where rownames are
HGNC genes (HUGO) and colnames are cell ids}

\item{sc_atac_data}{a single-cell ATAC-seq count matrix, where rownames are
cell ids and colnames are ids for euchromatine peaks (obtained from peak caller,
for ex. MACS2)}

\item{sc_atac_cell_names}{a matrix, containing description and annotation for
cell types observed in scATAC-seq count matrix. It should have rownames (cell
ids) that
match rownames of \code{sc_atac_data}}

\item{sc_atac_peaks_ann}{a matrix, containing description for euchromatine
peaks, presented in count matrix. It should have rownames (peak ids) that
match colnames of \code{sc_atac_data}}

\item{cell_types}{a vector, containing names for cell types, presented in
count matrix.}

\item{atac_bin_thrld}{a numeric value for accessibility threshold used for
primary binirization of peaks in scATAC-seq matrix.}
}
\value{
a data frame containg an imputed single-cell RNA-seq matrix
}
\description{
Epi-Impute, a computational tool for imputing scRNA-seq data from DNA accessibility data (scATAC-seq) from consistent cell-type populations.
}
